+++ qcdb/molecule/libmintsmolecule.py
@@ -37,7 +37,7 @@ import numpy as np
 import qcelemental as qcel
 
 from ..vecutil import *
-from ..exceptions import *
+from ..exceptions import IncompleteAtomError, ValidationError
 from .libmintscoordentry import NumberValue, VariableValue, CartesianEntry, ZMatrixEntry
 from .libmintspointgrp import SymmOps, similar, SymmetryOperation, PointGroup
